requirements.txt
black
pre-commit
pytest
numpy
pandas
matplotlib
serialtools
pyfirmata
